+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.133.0] - Not released
+### Added
+- Add a "description" meta-tag to the index.html.
+- Autocomplete for the user/group names in the post and comment inputs. When the
+  user types "@" and some text afterwards, the matched users/groups are shown
+  beneath the text input.
+### Changed
+- Switch to V3 server API (with _omittedCommentsOffset_ field and two comments
+  after the fold).
+### Fixed
+- Update SSI patterns in index.html to support dashes in groupnames.
+- Bug on iOS that blocks deletion of post's attachments.
+
 ## [1.132.0] - 2024-05-27
 ### Added
 - Ability to "unlock" and preview comments from the banned users and with reply